code cleanup, see description for details
renamed Models to small letter files.
removed (most) unnecessary print statements
moved docstrings to proper positions
