+++ Models/displayLevel.py
@@ -3,13 +3,11 @@ from mysql.connector import errorcode
 from Credentials import constants
 
 
-
-
-'''
-this function gets all the data stored in the Level table. 
-'''
 # to display level history saved in database
 def display():
+    """
+    This function gets all the data stored in the Level table.
+    """
     conn = mysql.connector.connect(host=constants.HOST,
                                    database=constants.DATABASE,
                                    user=constants.USER,
@@ -23,12 +21,12 @@ def display():
     return data
 
 
-'''
-delete() function takes in a variable which should be the map_id for the
-levels table. it deletes that row on the table. 
-    @param x        the map_id that determines the row to be deleted
-'''
 def delete(map_id):
+    """
+    delete() function takes in a variable which should be the map_id for the
+    levels table. it deletes that row on the table.
+        @param x        the map_id that determines the row to be deleted
+    """
     conn = mysql.connector.connect(host=constants.HOST,
                                    database=constants.DATABASE,
                                    user=constants.USER,
