+++ Models/processLogin.py
@@ -4,7 +4,7 @@ from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Length
 from flask_bcrypt import Bcrypt
 from datetime import datetime
-from .EditLevel import fetchPassword
+from .editLevel import fetchPassword
 
 bcrypt = Bcrypt()
 
@@ -14,8 +14,6 @@ class LoginForm(FlaskForm):
     submit = SubmitField('Log In')
     attempt = None
     incident_time = None
-    print(password)
-    
 
     def load(self):
         """
@@ -27,10 +25,8 @@ class LoginForm(FlaskForm):
             self.attempt = session['attempt']
             if self.attempt <= 0:
                 self.attempt = 0
-        # flash(f"\nNumber of attempts: {self.attempt}\n")
         if 'incident_time' in session:
             self.incident_time = int(session['incident_time'])
-            # flash(f"incident time: {int(self.incident_time)}, current_time: {int(datetime.now().timestamp())}, time diff: {int(datetime.now().timestamp()) - int(self.incident_time)}")
             if int(datetime.now().timestamp()) - int(self.incident_time) < 10:
                 self.password.render_kw = {'readonly': True}
             else:
@@ -49,13 +45,11 @@ class LoginForm(FlaskForm):
             hashed_input = bcrypt.generate_password_hash(self.password.data).decode('utf-8')
             db_pw = str(fetchPassword())[3:-4]
             if bcrypt.check_password_hash(hashed_input, db_pw):
-                #flash("Successful Login!", 'success')
                 return "Success"
             else:
                 if self.attempt < 1:
                     session['attempt'] = 0
                     if not 'incident_time' in session:
                         session['incident_time'] = datetime.now().timestamp()
-                    #flash(f'Too many incorrect logins incident"', 'danger')
                     return "Timeout"
                 return "Fail"
