+++ templates/command.html
@@ -31,9 +31,4 @@
 
   <h1>{{ Car_data }}</h1>
 
-  
-
-
-
-
 {% endblock %}
