+++ templates/dashboard.html
@@ -7,8 +7,6 @@
     {% endblock %}
 {% endblock %}
 
-
-
 {% block body %}
     <br>
     <div class="container rounded">
