+++ templates/index.html
@@ -16,8 +16,6 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.2/js/bootstrap.min.js" integrity="sha384-alpBpkh1PFOepccYVYDB4do5UnbKysX5WZXm3XxPqe5iKTfUKjNkCk9SaVuEZflJ" crossorigin="anonymous"></script>
     {% endblock %}
 
-    
-
 {% endblock %}
 
 {% block body %}
@@ -81,12 +79,7 @@
                 
             </div>
         </div>
-        
-
         <br>
-
-        
-
         <div id="control-div" class="container-fluid justify-content-center text-center">
             <h2> Car Control Station </h2>
             <div id="blocklyDiv" style="height: 320px; width: auto;"></div>
@@ -96,8 +89,6 @@
         <br>
       </div>
       <div class="col">
-            
-            
             <!-- Modal -->
             <div class="modal fade-modal-sm" id="exampleModalCenter" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
                 <div class="modal-dialog modal-dialog-centered modal-sm" role="document">
@@ -125,11 +116,7 @@
                 </div>
                 </div>
             </div>
-
       </div>
-   
-
-
 {% endblock %}
 
 {% block footer %}
